chore: Add getDateInTimezoneFromISO and getDateInTimezone functions to helpers
This commit adds the `getDateInTimezoneFromISO` and `getDateInTimezone` functions to the `helpers/index.ts` file. These functions allow for converting a given ISO date string or a `Date` object to a date in the 'America/Mexico_City' timezone. This addition enhances the functionality of the application by providing convenient methods for handling timezone conversions.