chore: Add Firebase configuration files
This commit adds the `.firebaserc` and `firebase.json` files to the repository. These files contain the necessary configuration for deploying the application to Firebase. The `.firebaserc` file specifies the default project as "alplanners", while the `firebase.json` file defines the functions configuration, including the source directory, codebase, ignore patterns, and predeploy script. This addition enables seamless deployment and ensures the correct setup of the Firebase environment.