chore: Update .gitignore to ignore compiled JavaScript files and TypeScript declaration files
This commit updates the .gitignore file to ignore the compiled JavaScript files in the "lib" directory and its subdirectories, as well as the TypeScript declaration files in the "typings" directory. Additionally, it ignores the "node_modules" directory and the "*.local" files. This change ensures that these files are not tracked by Git and improves the cleanliness of the repository.