refactor: Update User model schema to remove admin field
The User model schema in the src/models/User.ts file has been updated to remove the admin field. This change removes the admin property from the IUser interface and the userSchema, as it is no longer needed. The commit message follows the established convention of starting with "refactor" to indicate a code improvement or restructuring.